Trying to solve errors in CI
